requirements.txt

fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
openai==1.12.0
pinecone-client==3.0.2
httpx==0.24.1
motor==3.3.1
pymongo==4.6.0
python-multipart==0.0.5
email-validator==1.1.3
certifi==2024.2.2 
# //deployement  check